add lifespan and db script
